CTX-5356: Saving changes. All requests of task covered, need to improve the code a bit.
